feat: first algo works but not effective enough, quicksort on the way

+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: tale-fau <tale-fau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/27 15:45:48 by tale-fau          #+#    #+#             */
-/*   Updated: 2021/10/05 19:09:23 by tale-fau         ###   ########.fr       */
+/*   Updated: 2021/10/06 12:02:32 by tale-fau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_stack
 
 int		no_quotes(t_stack *a, t_stack *b, int ac, char **av);
 int		w_quotes(t_stack *a, t_stack *b, char *av);
-int		check_num(int num, t_stack *a, int len);
+int		check_num(long int num, t_stack *a, int len);
 int		error(int i);
 void	order(t_stack *a, t_stack *b);
 void	pa(t_stack *a, t_stack *b);
